[Refactor]: 조회수 로직 제거, 내림차순 정렬 추가
- 조회수 관련 로직 js 에서 제거
- 게시글 불러올 때 내림차순 정렬 추가
