+++ src/main/java/com/inconcert/domain/post/repository/InfoRepository.java
@@ -23,7 +23,8 @@ public interface InfoRepository extends JpaRepository<Post, Long> {
             "JOIN p.postCategory pc " +
             "JOIN pc.category c " +
             "JOIN p.user u " +
-            "WHERE c.title = 'info' ")
+            "WHERE c.title = 'info' " +
+            "ORDER BY p.createdAt DESC")
     List<PostDTO> findPostsByCategoryTitle(Pageable pageable);
 
     // info 포스트 상위 8개 불러오기
@@ -58,7 +59,8 @@ public interface InfoRepository extends JpaRepository<Post, Long> {
             "JOIN p.postCategory pc " +
             "JOIN pc.category c " +
             "JOIN p.user u " +
-            "WHERE c.title = 'info' AND pc.title = :postCategoryTitle")
+            "WHERE c.title = 'info' AND pc.title = :postCategoryTitle " +
+            "ORDER BY p.createdAt DESC")
     List<PostDTO> findPostsByPostCategoryTitle(@Param("postCategoryTitle") String postCategoryTitle);
 
     // /info/categoryTitle에서 게시물들 알맞게 불러오기
@@ -69,7 +71,8 @@ public interface InfoRepository extends JpaRepository<Post, Long> {
             "JOIN p.postCategory pc " +
             "JOIN pc.category c " +
             "JOIN p.user u " +
-            "WHERE c.title = 'info' AND pc.title = :postCategoryTitle")
+            "WHERE c.title = 'info' AND pc.title = :postCategoryTitle " +
+            "ORDER BY p.createdAt DESC")
     Page<PostDTO> findPostsByPostCategoryTitle(@Param("postCategoryTitle") String postCategoryTitle,
                                                Pageable pageable);
 
@@ -86,7 +89,8 @@ public interface InfoRepository extends JpaRepository<Post, Long> {
             "OR   (:type = 'title' AND p.title LIKE %:keyword%) " +
             "OR   (:type = 'content' AND p.content LIKE %:keyword%) " +
             "OR   (:type = 'author' AND u.nickname LIKE %:keyword%)) " +
-            "AND p.createdAt BETWEEN :startDate AND :endDate")
+            "AND p.createdAt BETWEEN :startDate AND :endDate " +
+            "ORDER BY p.createdAt DESC")
     Page<PostDTO> findByKeywordAndFilters(@Param("postCategoryTitle") String postCategoryTitle,
                                           @Param("keyword") String keyword,
                                           @Param("startDate") LocalDateTime startDate,
